Add tests to clarify selection of in, out, err
This is in response to #13, in which a user thought the err
channel should be set up to select for write like the out channel.
In actuality, out is the child's "in" so it is the only stream
which should be set up for write. in is the child's out and err is
the child's err, both readable by parent.